Quantum Optimization

BellScenario.optimize_measurement — Function

optimize_measurement(
    game :: BellGame,
    ρ_states :: Vector{<:States.AbstractDensityMatrix},
    PM :: PrepareAndMeasure
)

Perform semi-definite programming to find the optimal quantum measurement which maximizes the score for the specified set of quantum states ρ_states and BellGame.