Quantum Strategies

BellScenario.quantum_strategy — Function

Constructs a strategy matrix given quantum states and measurements. The supported scenarios include:

BlackBox scenarios

quantum_strategy(
    Π :: Observables.AbstractPOVM,
    ρ_states :: Vector{<:States.AbstractDensityMatrix}
) :: Strategy

LocalSignaling scenarios

quantum_strategy(
    Π :: Observables.AbstractPOVM,
    ρ_states :: Vector{<:States.AbstractDensityMatrix},
    scenario :: LocalSignaling
) :: Strategy

A DomainError is thrown if the provided states and measurements are not compatible with the specified scenario.