Black-box Scenarios

BellScenario.Scenario — Type

An abstract type to represent general black-box scenarios.

BellScenario.BlackBox — Type

BlackBox(num_in :: Int64, num_out :: Int64) <: Scenario

A black-box scenario considering a single device. A black-box is an uncharacterized device with a finite number of classical inputs and outputs.

A DomainError is throw if parameters num_in or num_out is less than 1.

BellScenario.Bipartite — Type

Bipartite(
    A :: Tuple{Int64, Int64},
    B :: Tuple{Int64, Int64};
    dits :: Int = 1,
    bidirectional :: Bool = false
) <: Scenario

A black-box scenario with two devices and possible communication between the devices. The keyword parameter dits describes the number of dits used for communication and bidirectional describes whether communication is simultaneous between each party.

BellScenario.PrepareAndMeasure — Type

PrepareAndMeasure(
    X :: Int,
    B :: Int,
    d :: Int,
)

A black-box scenario with X inputs, B outputs, and d-dits of communication from the input device to the output device.