Generating Vertices and Facets

BellScenario.LocalPolytope.generator_vertex — Function

generator_vertex(
    D :: DeterministicStrategy,
    scenario :: LocalSignaling
) :: DeterministicStrategy

Finds the generating vertex for the provided DeterministicStrategy. The generating vertex is the lexicographic normal form of D.

BellScenario.LocalPolytope.generator_facet — Function

generator_facet( BG :: BellGame, scenario :: LocalSignaling ) :: BellGame

Finds the generating facet for the provided BellGame. The generator is provided in lexicographic normal form. The generating facet is found recursively by an algorithm which sorts by lexicographic scores.