Vertex Enumeration

BellScenario.LocalPolytope.vertices — Function

vertices(
    PM :: PrepareAndMeasure;
    rep = "normalized" :: String
    rank_d_only = false :: Bool
) ::  Vector{Vector{Int64}}

Generates the deterministic strategies for the local polytope of PrepareAndMeasure scenarios. The rank_d_only keyword arg specifies whether to exclude vertices which use fewer dits of communication and thus have a matrix rank less than d.

Warning

The vertices computed in this method are vectorized directly from a strategy matrix by column-majorization. These vertices are distinct from those produced by older LocalPolytope.vertices() methods which are row-majorized. This functionality is a trial run of performance improvements to polytope computation.

BellScenario.LocalPolytope.num_vertices — Function

num_vertices( PM :: PrepareAndMeasure; rank_d_only = false :: Bool ) :: Int64

Counts the numbr of polytope vertices for the specified PrepareAndMeasure scenario. If rank_d_only = true, then only strategies using d-dits are counted.