Black-Box Games

BellScenario.AbstractGame — Type

Games represent a linear inequality.

BellScenario.Game — Type

Game(game::Matrix{T}, β::Real) <: AbstractMatrix{T}

A Game is played by a Matrix and β is the bounding or winning score of the game. A Game is matrix representation of a linear inequality. Each element is a linear scale factor for an element of a strategy matrix.

Type parameter T is typically an Int64 or Float64.

BellScenario.BellGame — Type

BellGame(game::Matrix{Int64}, β::Int64, scenario::Scenario)

A tight Bell inequality for the correlation polytope for scenario.

Base library extensions for BellGame:

Conversions between BellGames and other polytope facet represenations. A facet vector is a column-major vectorization of a game matrix, with the bound placed in the last element of the vector.

Facet (Vector{Int64}) -> BellGame

convert(
    ::Type{BellGame},
    facet::Vector{Int64},
    scenario::Scenario;
    rep = "normalized"::String
)

BellGame -> Facet (Vector{Int64})

convert(::Type{Vector{Int64}}, BG::BellGame; rep = "normalized" :: String)

BellGame's to XPORTA.IEQ

convert(::Type{IEQ}, bell_games::Vector{BellGame}; rep = "normalized" :: String)

XPORTA.IEQ to BellGame's

convert(
    ::Type{Vector{BellGame}},
    ieq::IEQ, scenario::Scenario;
    rep = "normalized" :: String
)