Black-box Strategies

BellScenario.AbstractStrategy — Type

A stochastic matrix which represents a map from input to output for a given Bell scenario.

Base library extensions for AbstractStrategy:

Two strategies can be multiplied together resulting in a new strategy, e.g. S1*S2 = S3.

*(S1::AbstractStrategy, S2::AbstractStrategy) :: Strategy

The chained strategies can describe a new black-box scenario, therefore, scenario parameter can also be passed to the multiplication operator.

*(S1::AbstractStrategy, S2::AbstractStrategy, scenario::Scenario) :: Strategy

BellScenario.Strategy — Type

A strategy matrix describing the statistical behavior of a black-box.

Strategy(conditionals :: Matrix{<:Real}) <: AbstractMatrix{Float64}

By default, the constructor creates a strategy for a 'BlackBox' scenario. However, a Scenario can be passed to the Strategy constructor.

Strategy(conditionals :: Matrix{<:Real}, scenario :: Scenario)

A DomainError is thrown if the provided Scenario does not match the dimension of the conditionals matrix.

BellScenario.DeterministicStrategy — Type

A strategy matrix describing the deterministic behavior of a black-box.

DeterministicStrategy(conditionals :: Matrix{Int64}) <: AbstractMatrix{Int64}

By default, the constructor creates a strategy for a 'BlackBox' scenario. However, a Scenario can be passed to the DeterministicStrategy constructor.

DeterministicStrategy(conditionals :: Matrix{Int}, scenario :: Scenario)

A DomainError is thrown if the provided Scenario does not match the dimension of the conditionals matrix.

A DomainError is thrown if the elements of conditionals are not 0 or 1.

Base library Extensions for DeterministicStrategy:

The product of two deterministic strategies is a DeterministicStrategy.

*(S1::DeterministicStrategy, S2::DeterministicStrategy) :: DeterministicStrategy

Deterministic strategies correspond to vertices of the local polytope. A vertex is simply represented by a Vector{Int64} and the rep argument specifies whether the vertex is in the "normalized" or "generalized" representation.

Vertex (Vector{Int64}) -> DeterministicStrategy

convert(
    ::Type{DeterministicStrategy},
    vertex  :: Vector{Int64},
    scenario :: Scenario;
    rep = "normalized" :: String
)

DeterministicStrategy -> Vertex (Vector{Int64})

convert(
    ::Type{Vector{Int64}},
    strategy :: DeterministicStrategy;
    rep = "normalized" :: String
)

BellScenario.is_deterministic — Function

is_deterministic( strategy :: AbstractMatrix  ) :: Bool

Returns true if all elements of strategy are either 0 or 1 and the matrix is a valid conditional probability distribution.

BellScenario.strategy_dims — Function

strategy_dims( scenario :: Scenario ) :: Tuple{Int, Int}

Returns the dimensions of the Conditionals matrix describing a Strategy for the Scenario at hand. Each Scenario, can place unique constraints on the matrix dimensions, therfore, separate methods are called for each concrete Scenario.

BellScenario.random_strategy — Function

random_strategy(
    num_inputs :: Int64,
    num_outputs :: Int64;
) :: Strategy

Constructs a randomized Strategy matrix. Zeros are inserted into the strategy to ensure that it does not closely resemble a uniform distribution.