Adjacency Decomposition

BellScenario.LocalPolytope.adjacency_decomposition — Function

adjacenecy_decomposition(vertices, BG_seed::BellGame, PM::PrepareAndMeasure; kwargs )

Given a polytpe represented by vertices, returns the complete set of canonical facets for prepare and measure scenario PM. The adjacencydecomposition algorithm requires a seeded vertex which is supplied with the `BGseed` argument. Facets are returned in the lexicographic normal form.

Returns a dictionary where the keys are canonical BellGames and the value is a dictionary with keys

"considered" => true, if the facet was considered in the algorithm.

"skipped" => true, if the facet was skipped (not considered).

"num_vertices" => number of vertices.

Keyword arguments kwargs

skip_games :: Vector{BellGame} - Optional list of games to skip.

max_vertices :: Int64 - Defaults to 100, the maximum number of vertices to allow in target facets.

BellScenario.LocalPolytope.adjacent_facets — Function

adjacent_facets( vertices, F::HalfSpace, PM::PrepareAndMeasure )

For the polytope represented by vertices, returns the canonical set of facets adjacent to F for the specified prepare and measure scenario PM.

F is assumed to ba a valid facet in the normalized subspace.

BellScenario.LocalPolytope.rotate_facet — Function

rotate_facet(F::HalfSpace, G::HalfSpace, xbar::Vector{Int})

Performs a rotation of facet F relative to non-included vertex xbar. Returns the rotated HalfSpace. F is a polytope facet, G is a facet of F and xbar is a vertex not contained on the boundary of F.